Fix argument type from int to long long (was n_blocks, is size_in_bytes now).

+++ src/XrdFileCache/XrdFileCacheFactory.cc
@@ -421,7 +421,7 @@ public:
 
    map_t fmap;
 
-   void checkFile (time_t iTime, const char* iPath,  int iNByte)
+   void checkFile (time_t iTime, const char* iPath,  long long iNByte)
    {
       if (nByteAccum < nByteReq || iTime < fmap.rbegin()->first)
       {
